güncelleme

+++ AracKiralamaOtomasyonu3/Forms/KiralayanForm.cs
@@ -1,4 +1,6 @@
 ﻿using System;
+using System.Drawing;
+using System.IO;
 using System.Linq;
 using System.Windows.Forms;
 using AracKiralamaOtomasyonu3.DAL;
@@ -9,14 +11,31 @@ namespace AracKiralamaOtomasyonu3
     public partial class KiralayanForm : Form
     {
         private int _kiralayanId;
+        private byte[] _aracResim;
 
         public KiralayanForm(int kiralayanId)
         {
             InitializeComponent();
             _kiralayanId = kiralayanId;
-       
         }
+        private void btnResimYukle_Click(object sender, EventArgs e)
+        {
+            using (OpenFileDialog openFileDialog = new OpenFileDialog())
+            {
+                openFileDialog.Filter = "Image Files|*.jpg;*.jpeg;*.png;*.bmp";
+                if (openFileDialog.ShowDialog() == DialogResult.OK)
+                {
+                    string filePath = openFileDialog.FileName;
+                    _aracResim = File.ReadAllBytes(filePath); // Resmi byte array olarak sakla
 
+                    using (var ms = new MemoryStream(_aracResim))
+                    {
+                        pbAracResim.Image = Image.FromStream(ms);
+                    }
+
+                }
+            }
+        }
         // Araç ekleme butonuna tıklama işlemi
         private void btnAracEkle_Click(object sender, EventArgs e)
         {
@@ -27,7 +46,10 @@ namespace AracKiralamaOtomasyonu3
                     Model = txtModel.Text,
                     Plaka = txtPlaka.Text,
                     Fiyat = decimal.Parse(txtFiyat.Text),
-                    BakimTarihi = dtpBakimTarihi.Value
+                    BakimTarihi = dtpBakimTarihi.Value,
+                    Resim = _aracResim
+
+
                 };
 
                 context.Araclar.Add(arac);
@@ -37,30 +59,37 @@ namespace AracKiralamaOtomasyonu3
                 ListeyiGuncelle();
             }
         }
-      
-
-        // Araç silme butonuna tıklama işlemi
         private void btnAracSil_Click(object sender, EventArgs e)
         {
             using (var context = new AracKiralamaContext())
             {
-                var plaka = txtPlaka.Text;
-                var arac = context.Araclar.FirstOrDefault(a => a.Plaka == plaka);
+                var plaka = txtPlaka.Text.Trim().ToUpper();
+
+                if (string.IsNullOrWhiteSpace(plaka))
+                {
+                    MessageBox.Show("Lütfen silmek istediğiniz aracın plaka alanını doldurun!", "Uyarı", MessageBoxButtons.OK, MessageBoxIcon.Warning);
+                    return;
+                }
+
+                var arac = context.Araclar.FirstOrDefault(a => a.Plaka.ToUpper() == plaka); // Veritabanında plakayı büyük harf olarak karşılaştırır
 
                 if (arac != null)
                 {
                     context.Araclar.Remove(arac);
                     context.SaveChanges();
                     MessageBox.Show("Araç başarıyla silindi!");
+                    ListeyiGuncelle(); // Silme sonrası güncelle
                 }
                 else
                 {
-                    MessageBox.Show("Araç bulunamadı!");
+                    MessageBox.Show("Araç bulunamadı veya plaka hatalı!", "Hata", MessageBoxButtons.OK, MessageBoxIcon.Error);
                 }
             }
-
-            ListeyiGuncelle();
         }
+
+
+
+
 
         // Kiralanan araçları listele
         private void btnKiralananAraclar_Click(object sender, EventArgs e)
@@ -104,9 +133,11 @@ namespace AracKiralamaOtomasyonu3
             ListeyiGuncelle();
         }
 
+
+        // txtModel değişikliği algılama
         private void txtModel_TextChanged(object sender, EventArgs e)
         {
-
+            // İlgili işlem yapılabilir.
         }
     }
 }
